+++ src/ocamlorg_web/lib/handlers/page_handler.ml
@@ -180,11 +180,12 @@ let industrial_users req =
     (Industrial_users_template.render users)
   |> Dream.html
 
-let academic _req =
+let academic req =
+  let users = with_lang req Ood.Academic_institution.all () in
   Page_layout_template.render
     ~title:"Academic Users of OCaml"
     ~description:"OCaml usage in the academic world."
-    (Academic_template.render Ood.Academic_institution.all)
+    (Academic_template.render users)
   |> Dream.html
 
 let what_is_ocaml _req =
@@ -286,15 +287,20 @@ let releases _req =
   Page_layout_template.render
     ~title:"Releases"
     ~description:"???"
-    (Releases_template.render ())
+    (Releases_template.render Ood.Release.all)
   |> Dream.html
 
-let release _req =
-  Page_layout_template.render
-    ~title:"Release"
-    ~description:"???"
-    (Release_template.render ())
-  |> Dream.html
+let release req =
+  let version = Dream.param "id" req in
+  match Ood.Release.get_by_version version with
+  | Some release ->
+    Page_layout_template.render
+      ~title:"Release"
+      ~description:"???"
+      (Release_template.render release)
+    |> Dream.html
+  | None ->
+    not_found req
 
 let tutorials _req =
   Page_layout_template.render